+++ src/cpptools.ts
@@ -340,7 +340,9 @@ export class CppConfigurationProvider implements cpptools.CustomConfigurationPro
     /** Our name visible to cpptools */
     readonly name = 'CMake Tools';
     /** Our extension ID, visible to cpptools */
-    readonly extensionId = 'ms-vscode.cmake-tools';
+    get extensionId(): string {
+        return util.thisExtensionId();
+    }
     /**
      * This value determines if we need to show the user an error message about missing compilers. When an update succeeds
      * without missing any compilers, we set this to `true`, otherwise `false`.